Move interface definitions next to entities that use them
