+++ src/Entity/RegistrantInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\rng;
+namespace Drupal\rng\Entity;
 use Drupal\Core\Entity\ContentEntityInterface;
 use Drupal\Core\Entity\EntityInterface;
