+++ src/Entity/RegistrationTypeInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\rng;
+namespace Drupal\rng\Entity;
 
 use Drupal\Core\Config\Entity\ConfigEntityInterface;
