+++ src/Entity/RuleComponent.php
@@ -3,10 +3,8 @@
 namespace Drupal\rng\Entity;
 
 use Drupal\Core\Entity\ContentEntityBase;
-use Drupal\rng\RuleComponentInterface;
 use Drupal\Core\Entity\EntityTypeInterface;
 use Drupal\Core\Field\BaseFieldDefinition;
-use Drupal\rng\RuleInterface;
 
 /**
  * Defines a event rule plugin instance entity: a condition or action.
